Refactor Configuration-to-Setting interface
A Configuration must provide two interfaces, one an inherent property
of the problet domain, the other to satisfy a design goal of this
library.

Fundamentally, a Configuration is a container for a collection of
Settings.  Therefore, it needs to provide access to them.

This library aims to be a "drop-in" configuration module.  Current
software practice grows an application organically through short
iterations rather than exhaustive planning.  Persistent settings may
not be requested until the application is minimally viable.  However,
the best time to integrate settings is up-front as updating requires
changing the call to each tracked variable, as well as how the mapping
between the call, UI, and permanent storage happens.  In such a
situation, it'd be convenient to simply replace "my_variable" with
"my_configuration.my_variable" and have the Setting itself manage the
mapping.

The two interface requirements suggest the following forms:

- Configuration[<setting>] to access the underlying Setting object
- Configuration.<setting> to access a setting value

The Setting object has a value, with a default, as well as a getter
and setter for managing the UI.

It must be observed that overriding attribute access for the Setting
value presents a conflict with Configuration methods.  This is an
unfortunate sacrifice of beauty and symmetry at the alter of
practicality[1].

[1] It is possible to override methods which shadow Settings in order
to accommodate both functionality.

See:
- https://stackoverflow.com/a/46312533
- https://stackoverflow.com/a/5079766