Implement 'get' method
Previously, the only way to bulk update an external component was to
either call 'read' or to manually iterate through the Settings and
call their getters.  The 'get' method allows users to update the
Configuration to match what end-users may have toggled in a settings
dialog without requiring a disk operation or knowing the internals of
Nostalgic.