Fix premature shell exit
Using 'set -e' caused the shell to exit on tab completion.  The 'set'
command is used to set shell attributes (when options are supplied).
The '-e' option will

> "Exit immediately if a pipeline (*note Pipelines::), which may
> consist of a single simple command (*note Simple Commands::), a
> list (*note Lists::), or a compound command (*note Compound
> Commands::) returns a non-zero status."

Recall that "non-zero status" is UNIX-speak for "error".  When "set
-e" is enabled and a command fails, *any* command, the shell will
exit.  This includes tab completion.  For this reason, "set -e"
shouldn't be used in interactive scripts.