Fix assignment bug
Implement `__setattr__`.

Without `__setattr__`, the value was being set directly on the
Configuration object and not in the Setting.  Because `__getattr__`
was implemented, there was an asymmetry in data access that made it
appear like the value was stored as expected.  The tests did not catch
the debug because they used either the default value or a custom
getter that didn't use the Setting object.  There is now an explicit
test to check the Setting object directly for the value.